Include peter's changes to classlist3

+++ lib/WeBWorK/DB.pm
@@ -1262,7 +1262,11 @@ sub deleteGlobalUserAchievement {
 	# userAchievementID can be undefined if being called from this package
 	my $U = caller eq __PACKAGE__ ? "!" : "";
 	my ($self, $userID) = shift->checkArgs(\@_, "user_id$U");
-	return $self->{global_user_achievement}->delete($userID);
+	if ($self->{global_user_achievement}){
+		return $self->{global_user_achievement}->delete($userID);
+	} else {
+		return 0;
+	}
 }
